Fixed Purification math. division
Пофиксил дележку в формуле Очищения, разобрался, что за лабуда с типами была. Вопрос баланса все еще открыт, но механика работает - и ладно.

+++ Blade&Kami/CombatLogic.h
@@ -9,7 +9,7 @@ public:
     static void processPlayerAction(Player& player, Enemy& enemy, int action);
     static void processEnemyAction(Player& player, Enemy& enemy);
 
-    static float calculatePurificationChance(Player& player, Enemy& enemy);
+    static double calculatePurificationChance(Player& player, Enemy& enemy);
     static bool calculateHit(float attackerAccuracy, float targetEvasion);
     static int calculateDamage(int minDmg, int maxDmg);
 };
